clean up lcd init process
this splits the 4-bit init from the lcd settings etc (expect for the
font and lines as this should only be done once according to the
datasheet)

now the 4 bit init sets the mode to 8 bit a few times to ensure the lcd
nibble order is synced on init, once in sync we switch to 4 bit mode and
init the lines and font; this makes the init much more reliable and
ensures we don't get our nibbles (4 bits) out of sync when reloading the
driver, hence we run this on every lcd_init regardless of hw_reset
state.

now during the lcd_init we power cycle the lcd (or we will when the power
switch is ready, but no harm in just doing this in the code so we can
debug it) only if hw_reset is defined; also if hw_reset is defined we
clear the screen etc to put it into a known state since it probably only
just got power; if hw_reset is not defined we just home the cursor so
at we insmod with the cursor in a known position.